feat: externalized RawArticle schema, types and utils
