+++ schemas/article/Article.ts
@@ -1,4 +1,5 @@
 import { z } from 'zod'
+import { RawArticleSchema } from '@/schemas/article/RawArticle'
 
 export const ArticleSchema = z.object({
   id: z.number().optional(),
@@ -16,29 +17,9 @@ export const ArticleSchema = z.object({
   category_id: z.number().optional(),
 })
 
-export const RawArticleSchema = z.object({
-  article_id: z.string().optional(),
-  article_category_id: z.number().optional(),
-  article_name: z.string().default('Article-XY'),
-  article_code: z.string().optional(),
-  article_eanCode: z.string().optional(),
-  article_taxRate: z.string().optional(),
-  article_unit: z.string().optional(),
-  article_net_purchacePrice: z.string().optional(),
-  article_net_sellingPrice: z.string().optional(),
-  article_gross_sellingPrice: z.string().optional(),
-  article_stock: z.string().optional(),
-  article_minStock: z.string(),
-  article_comment: z.string().optional(),
-  article_billReference: z.string(),
-  article_negativeStock: z.boolean().optional().catch(false),
-  article_stockStatus: z.number().optional().catch(0),
-})
-
 export const RawArticlesSchema = z.object({
   articles: z.array(RawArticleSchema),
 })
 
 export type Article = z.infer<typeof ArticleSchema>
-export type RawArticle = z.infer<typeof RawArticleSchema>
 export type RawArticles = z.infer<typeof RawArticlesSchema>
